requirements.txt
langchain==0.3.14
langchain-core==0.3.29
langchain_community==0.3.14
langgraph==0.2.48
langgraph-checkpoint==2.0.4
python-dotenv==1.0.1
langchain-anthropic==0.2.1
langchain-openai==0.3.0
langchain-ollama==0.2.0
playwright==1.47.0
libsass==0.23.0
openai==1.59.7
cohere==5.10.0
langchain-cohere==0.3.0
chromadb==0.4.21
todoist-api-python==2.1.7
termcolor==2.4.0
lxml==5.3.0
sounddevice==0.5.0
soundfile==0.12.1
keyboard==0.13.5
pygments==2.18.0
rich==13.9.2
vue-lexer==0.0.4
requests==2.32.3
click==8.1.7
pytest==8.3.4
pytest-cov==6.0.0
pyright==1.1.390
ruff==0.8.2
httpx==0.27.2
questionary==2.1.0
crawl4ai==0.3.744
setuptools==75.8.0